refactor: refactor blank line before first member and abbreviation
