+++ src/main/java/racingcar/exception/EmptyInputException.java
@@ -1,6 +1,7 @@
 package racingcar.exception;
 
 public class EmptyInputException extends IllegalArgumentException {
+
     public EmptyInputException() {
         super("빈 입력값은 불가합니다.");
     }
