+++ src/main/java/racingcar/exception/TheCarDoesntStartException.java
@@ -1,6 +1,7 @@
 package racingcar.exception;
 
 public class TheCarDoesntStartException extends IllegalArgumentException {
+
     public TheCarDoesntStartException() {
         super("자동차가 출발하지 않았습니다.");
     }
