+++ src/main/java/racingcar/service/GameService.java
@@ -1,10 +1,10 @@
 package racingcar.service;
 
 import static java.util.Comparator.*;
-import static racingcar.util.ConstNumber.CAR_IS_NOT_STARTED_YET;
-import static racingcar.util.ConstNumber.CAR_MOVE_CONDITION;
-import static racingcar.util.ConstNumber.RANDOM_NUMBER_END_WITH;
-import static racingcar.util.ConstNumber.RANDOM_NUMBER_START_WITH;
+import static racingcar.util.MoveCarNumberConstant.CAR_IS_NOT_STARTED_YET;
+import static racingcar.util.MoveCarNumberConstant.CAR_MOVE_CONDITION;
+import static racingcar.util.MoveCarNumberConstant.RANDOM_NUMBER_END_WITH;
+import static racingcar.util.MoveCarNumberConstant.RANDOM_NUMBER_START_WITH;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,7 +14,7 @@ import racingcar.model.Car;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class GameService {
-    // TODO : 상수화 리팩토링
+
     public boolean isMoving() {
         int carCondition = randomNumber();
         return carCondition >= CAR_MOVE_CONDITION.getValue();
