+++ src/main/java/racingcar/util/OutputMessageConstant.java
@@ -1,6 +1,7 @@
 package racingcar.util;
 
-public enum ConstOutputMessage {
+public enum OutputMessageConstant {
+
     INPUT_CAR_NAME("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
     INPUT_TURN_COUNT("시도할 횟수는 몇 회인가요?"),
     EXECUTION_RESULT("실행 결과"),
@@ -8,7 +9,7 @@ public enum ConstOutputMessage {
 
     private String message;
 
-    ConstOutputMessage(String message) {
+    OutputMessageConstant(String message) {
         this.message = message;
     }
